fix user pop up nav, detail add to cart

+++ app/src/main/java/com/betulantep/bootcampgraduationproject/ui/detail/DetailViewModel.kt
@@ -1,10 +1,14 @@
 package com.betulantep.bootcampgraduationproject.ui.detail
 
+import android.view.View
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.betulantep.bootcampgraduationproject.R
 import com.betulantep.bootcampgraduationproject.data.entity.Basket
 import com.betulantep.bootcampgraduationproject.data.entity.Food
 import com.betulantep.bootcampgraduationproject.data.repo.BasketRepository
+import com.betulantep.bootcampgraduationproject.utils.showSnackBar
+import com.betulantep.bootcampgraduationproject.utils.showToast
 import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
@@ -26,20 +30,26 @@ class DetailViewModel @Inject constructor(var basketRepo: BasketRepository) : Vi
         basketRepo.getFoodQuantity()
 
     }
-    fun clickedAddToCart(food: Food,quantity: Int) {
-        if(!basketFoodList.value.isNullOrEmpty()){
-            for(basketFood in basketFoodList.value!!){
-                if(basketFood.basket_food_name == food.foodName){
-                    basketRepo.deleteFoodBasket(basketFood.basket_food_id,userName)
+    fun clickedAddToCart(view:View, food: Food,quantity: Int) {
+        if(quantity == 0){
+            showToast(view.context, R.string.lutfen_urun_adeti_seçiniz)
+        }else{
+            loadAllFoodBasket()//detaydan çıkmadan tekrar ürün eklerse 2 kere eklenmesin diye
+            if(!basketFoodList.value.isNullOrEmpty()){
+                for(basketFood in basketFoodList.value!!){
+                    if(basketFood.basket_food_name == food.foodName){
+                        basketRepo.deleteFoodBasket(basketFood.basket_food_id,userName)
+                    }
                 }
             }
+            basketRepo.addFoodBasket(
+                food.foodName,
+                food.foodImageName,
+                food.foodPrice,
+                quantity,
+                userName
+            )
+            showSnackBar(view, R.string.urun_sepete_eklendi)
         }
-        basketRepo.addFoodBasket(
-            food.foodName,
-            food.foodImageName,
-            food.foodPrice,
-            quantity,
-            userName
-        )
     }
 }
